[WIP] Implement SystemTray::set_icon to change the icon at runtime
Draft PR to solve https://github.com/tauri-apps/tao/issues/64

- [x] Linux
- [ ] Windows
- [ ] macOS

Submitting this early since I have the feeling that I'm making the code
a lot uglier. I'm new to Rust in terms of actual contributions so I'd
like a first pass review to see if I'm approaching this the right way.

I decided to add an `instance` field to `SystemTray` so that it can hold
the platform-specific system tray instance, in the case of Linux that's
an `AppIndicator`. This is so that later on we can access that instance
during runtime and use its `set_icon` method.

Later on this can be extended to `set_items` to also solve https://github.com/tauri-apps/tao/issues/63