Add Linux method to use existing GTK window
Fix #925

The existing `Window::new()` code can be roughly split into two parts:
applying generic attributes to the window and connecting the window to
the event loop. These can be fairly cleanly separated, so I did that -
the new `new_from_gtk_window` method contains the latter, where `new`
now just contains the former and calls `new_from_gtk_window` for the
latter.

There are a few attributes that affect both the former and latter parts,
and I presume in such cases the former is required for the latter to
function properly. It's up to the caller to make sure that those
prerequisites are met, but I'd like to add specific guidance to the new
attribute struct documentation if this approach looks reasonable.
(regarding `transparent`, `cursor_moved`, `fullscreen`, etc)